Fix bug where number of bytes transmitted was coming out negative, leading to a huge energy used number. Make sure that large manifest constants are treated as unsigned. Add log points EVENT_WAITING_POWER, EVENT_ENERGY_REQUIRED_UWH, EVENT_ENERGY_REQUIRED_TOTAL_UWH, EVENT_ENERGY_USED_UWH, rename EVENT_STACK_LEFT to EVENT_STACK_MIN_LEFT (and similar renaming to the associated debug function) and remove EVENT_BME280_ERROR as it was unused.  Log APPLICATION_LOG_VERSION becomes 5.

+++ source/actions/act_voltages.cpp
@@ -97,9 +97,9 @@ int getVPrimaryMV()
 }
 
 // Get the [optimistic] estimate of energy available.
-unsigned int getEnergyOptimisticNWH()
+uint64_t getEnergyOptimisticNWH()
 {
-    unsigned int energyNWH = 0;
+    uint64_t energyNWH = 0;
 
     // The idea here is to make sure that the
     // processor tries to do things, even if it
